fix: fix heartbeat startup/shutdown timing bug
The heartbeat code sets a timeout to start the heartbeat then sets
an interval to perform the heartbeat.

If the node is shut down after the timeout is set but before it
fires, the interval is set and the process never ends.

This PR refactors the heatbeat shutdown to clear the timeout as well
as the interval.