fix: use typed event emitter classes
The typed `EventEmitter` class from `@libp2p/interface/events` has
been renamed `TypedEventEmitter` along with a `TypedEventTarget`
interface.

The old name is still exported so it's a backwards compatible change
but it will be removed in a future version of `@libp2p/interface`
so this change future-proofs `@chainsafe/libp2p-gossipsub` against
that upcoming release.