fix: remove abortable iterator
AbortableSource is slow because it races promises against every chunk
causing extra async work.

It's only really necessary if we're going to pass the source off to
another component.

Here we don't do that so it's simpler to just add a listener for the
abort event and close the stream.