feat: add service capabilities and dependencies definition
The latest libp2p release allows services to declare their
capabilities (what they provide to the node) and their
dependencies (the capabilities they require other services to
provide for the config to be "valid").

Adds the `serviceCapabilities` symbol property that says Gossipsub
provides the `@libp2p/pubsub` capability and also that it depends
on another service providing the `@libp2p/identify` capability -
this is required because Gossipsub uses network topologies to
discover peers which requires the Identify protocol.